Remove ThumbnailVerticle
