+++ src/main/resources/beanRefContext.xml
@@ -73,23 +73,13 @@
     <constructor-arg value="${omero.pixeldata.max_tile_length:2048}" />
     <constructor-arg ref="imageRegionPixelsService"/>
     <constructor-arg ref="zarr-utils"/>
+    <constructor-arg ref="iscale" />
   </bean>
 
   <bean id="omero-ms-shape-mask-verticle"
         class="com.glencoesoftware.omero.ms.image.region.ShapeMaskVerticle"
         scope="prototype">
     <constructor-arg ref="zarr-utils" />
-  </bean>
-  
-  <bean id="omero-ms-thumbnail-verticle"
-        class="com.glencoesoftware.omero.ms.image.region.ThumbnailVerticle"
-        scope="prototype">
-    <constructor-arg ref="internal-ome.api.ICompress" />
-    <constructor-arg ref="lutProvider" />
-    <constructor-arg value="${omero.pixeldata.max_tile_length:2048}" />
-    <constructor-arg ref="imageRegionPixelsService" />
-    <constructor-arg ref="zarr-utils" />
-    <constructor-arg ref="iscale" />
   </bean>
   
   <bean id="rendering-utils"
